qualtdict: Generate Variable Dictionaries and Labelled Survey Data for 'Qualtrics' Surveys

Generates Variable Dictionaries from 'Qualtrics' https://www.qualtrics.com/about/ survey metadata and applies those dictionaries to create Labelled Survey Data. This package is a metadata and labelled-export companion to 'qualtRics', which provides access to 'Qualtrics' survey data and metadata via the 'Qualtrics' API https://api.qualtrics.com/.

See also

Author

Maintainer: Yuhao Lin yuhao.lin@kcl.ac.uk (ORCID)

Authors:

Yuhao Lin yuhao.lin@kcl.ac.uk (ORCID)